BUG: Samples are a arithmetic type

+++ src/apps/scdetect/template.cpp
@@ -228,7 +228,7 @@ namespace template_detail {
 
 bool XCorr(const double *tr1, const int size_tr1, const double *tr2,
            const int size_tr2, const double sampling_freq,
-           const double max_lag_samples, Template::MatchResultPtr result) {
+           const int max_lag_samples, Template::MatchResultPtr result) {
 
   /*
    * Pearson correlation coefficient for time series X and Y of length n
